Add DIP upload to SS script
Uploads a local DIP to a DIP storage location in an SS instance.
Requires access to a pipeline's currently processing location path (the
shared path), to move the DIP folder in there and send a requests to the
Storage Service to process that DIP and create a relationship with the
AIP from where it was created.

Add METS type argument to `create_dip` script to choose between a
generated METS file that matches AtoM requirements for DIP upload and
copying the AIP's METS file directly, like the Storage Service does, to
be able to move the DIP to the SS. Defaults to AtoM METS for backwards
compatibility.

Allow to upload DIPs after creation in `create_dips_job`. Accept
different subsets of parameters and optionally upload the created DIPs
to AtoM or the Storage Service. Extend this script to make use of its
tracking functionality to create DIPs from all AIPs in a SS location.
Optionally delete the created DIPs after upload.

Use AIP folder name for DIP folder on creation to match what's currently
being done for DIPs stored in the SS.

+++ aips/create_dip.py
@@ -25,13 +25,13 @@ import amclient
 import metsrw
 
 THIS_DIR = os.path.abspath(os.path.dirname(__file__))
-LOGGER = logging.getLogger("create_dip")
+LOGGER = logging.getLogger("dip_workflow")
 
 
 def setup_logger(log_file, log_level="INFO"):
     """Configures the logger to output to console and log file"""
     if not log_file:
-        log_file = os.path.join(THIS_DIR, "create_dip.log")
+        log_file = os.path.join(THIS_DIR, "dip_workflow.log")
 
     CONFIG = {
         "version": 1,
@@ -53,14 +53,14 @@ def setup_logger(log_file, log_level="INFO"):
             },
         },
         "loggers": {
-            "create_dip": {"level": log_level, "handlers": ["console", "file"]}
+            "dip_workflow": {"level": log_level, "handlers": ["console", "file"]}
         },
     }
 
     logging.config.dictConfig(CONFIG)
 
 
-def main(ss_url, ss_user, ss_api_key, aip_uuid, tmp_dir, output_dir):
+def main(ss_url, ss_user, ss_api_key, aip_uuid, tmp_dir, output_dir, mets_type="atom"):
     LOGGER.info("Starting DIP creation from AIP: %s", aip_uuid)
 
     if not os.path.isdir(tmp_dir):
@@ -105,7 +105,7 @@ def main(ss_url, ss_user, ss_api_key, aip_uuid, tmp_dir, output_dir):
         return 5
 
     LOGGER.info("Creating DIP")
-    dip_dir = create_dip(aip_dir, aip_uuid, output_dir)
+    dip_dir = create_dip(aip_dir, aip_uuid, output_dir, mets_type)
 
     if not dip_dir:
         LOGGER.error("Unable to create DIP")
@@ -115,6 +115,8 @@ def main(ss_url, ss_user, ss_api_key, aip_uuid, tmp_dir, output_dir):
     shutil.rmtree(tmp_dir)
 
     LOGGER.info("DIP created in: %s", dip_dir)
+
+    return dip_dir
 
 
 def extract_aip(aip_file, aip_uuid, tmp_dir):
@@ -156,7 +158,7 @@ def extract_aip(aip_file, aip_uuid, tmp_dir):
     return extract_aip(extracted_entry, aip_uuid, tmp_dir)
 
 
-def create_dip(aip_dir, aip_uuid, output_dir):
+def create_dip(aip_dir, aip_uuid, output_dir, mets_type):
     """
     Creates a DIP from an uncompressed AIP.
 
@@ -165,8 +167,9 @@ def create_dip(aip_dir, aip_uuid, output_dir):
     :param str output_dir: absolute path to a directory to place the DIP
     :returns: absolute path to the created DIP folder
     """
-    aip_name = os.path.basename(aip_dir)[:-37]
-    dip_dir = os.path.join(output_dir, "{}_{}_DIP".format(aip_name, aip_uuid))
+    aip_dir_name = os.path.basename(aip_dir)
+    aip_name = aip_dir_name[:-37]
+    dip_dir = os.path.join(output_dir, aip_dir_name)
     objects_dir = os.path.join(dip_dir, "objects")
     to_zip_dir = os.path.join(objects_dir, aip_name)
 
@@ -261,46 +264,54 @@ def create_dip(aip_dir, aip_uuid, output_dir):
         timestamp = int(fslastmodified) // 1000
         os.utime(dip_file_path, (timestamp, timestamp))
 
-    # Modify METS file for DIP
-    objects_entry = None
-    for fsentry in fsentries:
-        # Do not delete AIP entry
-        if (
-            fsentry.label == os.path.basename(aip_dir)
-            and fsentry.type.lower() == "directory"
-        ):
-            continue
+    # Modify or copy METS file for DIP based on mets_type argument
+    dip_mets_file = os.path.join(dip_dir, "METS.{}.xml".format(aip_uuid))
+    if mets_type == "atom":
+        LOGGER.info("Creating DIP METS file for AtoM upload.")
+        objects_entry = None
+        for fsentry in fsentries:
+            # Do not delete AIP entry
+            if (
+                fsentry.label == os.path.basename(aip_dir)
+                and fsentry.type.lower() == "directory"
+            ):
+                continue
 
-        # Do not delete objects entry and save it for parenting
-        if fsentry.label == "objects" and fsentry.type.lower() == "directory":
-            objects_entry = fsentry
-            continue
+            # Do not delete objects entry and save it for parenting
+            if fsentry.label == "objects" and fsentry.type.lower() == "directory":
+                objects_entry = fsentry
+                continue
 
-        # Delete all the others
-        mets.remove_entry(fsentry)
+            # Delete all the others
+            mets.remove_entry(fsentry)
 
-    if not objects_entry:
-        LOGGER.error("Could not find objects entry in METS file")
-        return
+        if not objects_entry:
+            LOGGER.error("Could not find objects entry in METS file")
+            return
 
-    # Create new entry for ZIP file
-    entry = metsrw.FSEntry(
-        label="{}.zip".format(aip_name),
-        path="objects/{}.zip".format(aip_name),
-        file_uuid=str(uuid.uuid4()),
-    )
+        # Create new entry for ZIP file
+        entry = metsrw.FSEntry(
+            label="{}.zip".format(aip_name),
+            path="objects/{}.zip".format(aip_name),
+            file_uuid=str(uuid.uuid4()),
+        )
 
-    # Add new entry to objects directory
-    objects_entry.add_child(entry)
+        # Add new entry to objects directory
+        objects_entry.add_child(entry)
 
-    # Create DIP METS file
-    LOGGER.info("Creating DIP METS file")
-    dip_mets_file = "{}/METS.{}.xml".format(dip_dir, aip_uuid)
-    try:
-        mets.write(dip_mets_file, fully_qualified=True, pretty_print=True)
-    except Exception:
-        LOGGER.error("Could not create DIP METS file")
-        return
+        # Create DIP METS file
+        try:
+            mets.write(dip_mets_file, fully_qualified=True, pretty_print=True)
+        except Exception:
+            LOGGER.error("Could not create DIP METS file")
+            return
+    else:
+        LOGGER.info("Copying AIP's METS file.")
+        try:
+            shutil.copy(to_zip_mets_file, dip_mets_file)
+        except Exception:
+            LOGGER.error("Could not create DIP METS file")
+            return
 
     # Compress to_zip_dir inside the DIP objects folder
     LOGGER.info("Compressing ZIP folder inside objects")
@@ -357,6 +368,12 @@ if __name__ == "__main__":
         help="Absolute path to the directory used to place the final DIP. Default: /tmp",
         default="/tmp",
     )
+    parser.add_argument(
+        "--mets-type",
+        choices=["atom", "storage-service"],
+        default="atom",
+        help="Generate METS file for AtoM upload or use the AIP's METS file for Storage Service upload. Default: atom.",
+    )
 
     # Logging
     parser.add_argument(
@@ -392,13 +409,20 @@ if __name__ == "__main__":
 
     setup_logger(args.log_file, log_level)
 
-    sys.exit(
-        main(
-            ss_url=args.ss_url,
-            ss_user=args.ss_user,
-            ss_api_key=args.ss_api_key,
-            aip_uuid=args.aip_uuid,
-            tmp_dir=args.tmp_dir,
-            output_dir=args.output_dir,
-        )
+    ret = main(
+        ss_url=args.ss_url,
+        ss_user=args.ss_user,
+        ss_api_key=args.ss_api_key,
+        aip_uuid=args.aip_uuid,
+        tmp_dir=args.tmp_dir,
+        output_dir=args.output_dir,
+        mets_type=args.mets_type,
     )
+
+    # The main function returns the DIP's path on success
+    # or an int higher than 0 if it fails. The scrip will
+    # always exit with an int, 0 on success.
+    if type(ret) != int:
+        ret = 0
+
+    sys.exit(ret)
